requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
aiosqlite==0.19.0
alembic==1.13.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
exceptiongroup==1.2.0
fastapi==0.104.1
frozenlist==1.4.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
iniconfig==2.0.0
Mako==1.3.0
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.26.2
packaging==23.2
pluggy==1.3.0
pydantic==2.5.2
pydantic_core==2.14.5
pytest==7.4.3
pytest-asyncio==0.23.2
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
yarl==1.9.4
asyncpg
gunicorn
python-dotenv~=1.0.0